Add Subscription as a concept on how to deal with Kafka topics of interest.
Consumer group instances are registered with a subscription, and the consumer group will consume all partitions of the topics that match the subscription.

Also, the subscription objects responds to topic changes using zookeeper watches. This means that a subscription can pick up on newly added topics and partitions.

We can also use the subscription concept to do regular cleaning of consumergroup metadata in zookeeper. E.g. we can remove claims and stored offsets for topics that a consumer group is no longer subscribed to.