Raises Kazoo exception: broker is down and in ISR
In certain conditions (i.e. an existing broker has recently joined the
cluster), a partition ISR list includes a broker that Kazoo
does not have information about (due to caching).  While this is rare, it is
exacerbated by restarting brokers and triggering topology changes.
There are also likely other (better) solutions to this; however this fix
is most expedient.